rename component files to ease importing
