+++ src/index.js
@@ -4,8 +4,8 @@ import ReactDOM from "react-dom";
 import { Provider } from "react-redux";
 
 // components
-import App from "./components/app/app";
-import ErrorBoundary from "./components/error-boundary/error-boundary";
+import App from "./components/app";
+import ErrorBoundary from "./components/error-boundary";
 
 // styles
 import "./index.css";
